requirements.txt
python-telegram-bot==20.6
nest-asyncio
emoji
flask
